refactor: extract query structs into their own modules
